refactor: rename logger and add telegram override
