+++ systems/scripts/handle_top_of_hour.py
@@ -15,7 +15,7 @@ from systems.scripts.get_window_data import get_wave_window_data_df
 from systems.scripts.kraken_utils import get_live_price, get_kraken_balance
 from systems.scripts.execution_handler import execute_buy, execute_sell
 from systems.scripts.ledger import Ledger
-from systems.utils.logger import addlog
+from systems.utils.addlog import addlog
 from systems.utils.path import find_project_root
 
 
@@ -240,14 +240,12 @@ def handle_top_of_hour(
                     verbose_state=True,
                 )
                 addlog(
-                    f"[LIVE] {ledger_name} | {tag} | {window_name} window",
+                    f"[LIVE] {ledger_name} | {tag} | {window_name} window\n"
+                    f"✅ Buy attempts: {buy_count} | Sells: {sell_count} | "
+                    f"Open Notes: {summary['open_notes']} | Realized Gain: ${summary['realized_gain']:.2f}",
                     verbose_int=1,
                     verbose_state=True,
-                )
-                addlog(
-                    f"✅ Buy attempts: {buy_count} | Sells: {sell_count} | Open Notes: {summary['open_notes']} | Realized Gain: ${summary['realized_gain']:.2f}",
-                    verbose_int=1,
-                    verbose_state=True,
+                    telegram=True,
                 )
 
             if not dry_run:
