+++ systems/scripts/trade_logic.py
@@ -5,7 +5,7 @@ from __future__ import annotations
 from typing import Dict
 
 from systems.scripts.ledger import Ledger
-from systems.utils.logger import addlog
+from systems.utils.addlog import addlog
 
 
 def maybe_buy(
